requirements.txt
django==4.0
django-bootstrap-v5==1.0.8
python-a2s==1.3.0
requests==2.27.1
django-environ==0.8.1